Create main.cpp
Here is an example I had to complete in the Udacity C++ Nano-Degree Program.  Here I had to create a class, the class would declare a string member variable with the name "name".  I also had to create a constructor for the class that initialized the variable "name" via a initializer list.  Lastly, I had to make the content of the class all "public".