Create main.cpp
This was an exercise I did in the Udacity C++ nano-degree program.  In the exercise I am supposed to finish create the required parts of the BankAccount class, utilizing public and private member variables/functions, accessors and mutators, and scope resolution.  In the end, the program is supposed to output the details of the bank account that was created.