Create main.cpp
This is an exercise that I finished in the Udacity C++ Nano-Degree program.  In this exercise I created a class with mutators and accessors, however one of the mutator and accessors methods involves the conversion between a cstring and a std::string.  This is because in the SetBrand method a std::string is passed, however brand name is a cstring.  In the GetBrand accessor method, the cstring is converted into a std::string as the method is supposed to return a std::string.