Create main.cpp
This is one of the exercises I have completed in the Udacity C++ Nano-Degree program I am doing.  In this exercise I create the Pyramid class, which calculates the volume of a pyramid with given dimensions.  This class also makes use of exception handling by using the "throw" clause, when it detects a invalid measurement for a dimension of the pyramid.