Create main.cpp
This is an exercise I had to complete in the Udacity C++ Nano-Degree program that I am enrolled in.  In this exercise I have to create a "Student" class with member variable name, grade, and GPA.  I have to make accessors and mutaters for these member variables, and use assert to test them.  I also tested the invariants in the mutater functions by utilizing exception handling.  In the end of the program, I use assertion to test if the exception handling was successful.