Create main.cpp
This is one of the exercises that I had to do in the Udacity C++ Nano-Degree program that I am enrolled in.  In this example, I created the String() function which converts the data in the Date class to return a string representing the date.  The purpose of the exercise that I did was to understand the purpose of abstraction, basically the user does not have to know how the String() function was implemented, they just call the String() function from the Date class interface.  I could have used an if-if else -else, blocks and still it would not impact how other users of the class would use the interface.