Create main.cpp
This is one of the exercises I had to do in the nano-degree program.  Here I completed the Sphere class, where the constructor gets a radius, and there are two accessors one for returning the value of the radius, and the other for returning the value of the volume.  The purpose of the exercise was for me to see how it would not matter, the manner in which the value of the volume is returned after calling the Volume() member function.  Instead of me doing the calculation of the volume before hand as seen as I did above, I could have just calculated the volume each time.  This basically was for me to see how abstraction works in C++.