Create main.cpp
This is one of the exercises that I had to do and completed, in the nano-degree program.  Here, I was given a pre-made sphere class, where I had to change some items in the class and add a mutator function for the radius.  I had to do this without loosing the ability to use an invariant.  To do this, I added another member function as private, for checking the radius range is valid, as well as updating the value of volume.  This new private member variable is called by the constructor and by the radius mutator member function that I also added to the class as part of the exercise.