Create main.cpp
This is one of the exercises I completed in the nano-degree program.  In this example, I have to refactor the Sphere class, such that there is only one function in the class, and that given a radius, I have to call the Volume method as the radius as an argument.  At the same time, I do not create any object of type sphere.  This is accomplished by creating volume function that is static, and directly calling it.  The purpose of this exercise was to make me understand, a static member function can be called even if there is no object of the type of class.