Create main.cpp
This was one of the exercises I completed in the C++ nano-degree program I am enrolled in.  In this exercise, I was given some code, where I had to add another member variable in the parent class as well as doing other tasks as described further in this commit comment.  I then modified the Print() function in the parent class so that it displays the data of the new member variable that I added to the parent class.  I then initialized the member variable in the pre-instantiated Car object.  I also added a statement where the class member variable that I created had its data accessed directly from the main() function, when printing information about it only.  

I then crated another child class "Truck" whose parent class was "Vehicle".  I instantiated a Truck object, and initialized its inherited member variables plus it's own member variable I created called "loadedWithCargo".  I then called the print function, to display the info about the object I created.  Then I used a logic statement, where the program will decide on whether to print out a statement in regards to whether there is cargo on the Truck object or not.  

The purpose of this exercise was to get me familiar with the concept of Inheritance in C++.