Create main.cpp
This is one of the exercises that I did in the Udacity nano-degree program that I enrolled in.  In this exercise, I was given requirements about what my code is to do.  I made 3 classes, the Animal class (parent) and the Snake and Cat classes (child classes).  I also added a unique member variable to each of the child classes.  I also added a MakeSound() function to each of the child classes, where it would print out to the console the description of the object and the sound it makes.  In the main function, I instantiated an instance of each of the two child classes, then initialized the member variables of the child's and the member variables that it inherited.  I then called the MakeSound() function for each of the two classes.