Create main.cpp
This is an exercise I did in the Udacity nano-degree program that I am enrolled in.  Here I created two classes, where one of them uses compositional relationship with the other, basically a "has a" relationship, rather than the "is a" relationship that is seen in inheritance.  I also used the protected access modifier, which required the use of setters and getters in either of the two classes in order for the program to work.  The main purpose of the exercise was for me to understand "composition", and get more practice on access modifiers.