Create main.cpp
This is one of the exercises that I completed in the enrolled Udacity nano-degree program.  Here I was tasked to use my knowledge of method/function overloading polymorphism and what I learned on how how overloading polymorphism is implemented in C++, is similar or same in a sense in how it is done in Java.  Here I had to overload the function hello(), and create three different classes to use their objects as a parameter in the overloaded functions.