Create main.cpp
This is one of the exercises that I did in the Udactiy Nano-Degree program I am enrolled in.  In this exercise, I have the task of creating a "Point" class, and overloading the "+" operator, to add 2 Point objects together, to get a third Point object.  The calculation of the third object is done via adding the two x values (one from the base object (left of operator) and the addend Point object).  The same is done for calculating the Y value of the third Point object.  The purpose of this exercise was to reinforced what I learned on how Operator Overloading is done in C++.