Create main.cpp
This is one of the exercises I did in the Udacity nano-degree program I am enrolled in.  This exercise involved me creating two virtual functions, and implementing them in the child classes.  This was to deepen my knowledge into C++ on how virtual functions work.  They are basically interfaces, and are similar in a sense of pupouse to "interface" files in Java.