Create main.cpp
This is one of the exercises that I did in the Udacity nano-degree program.  Here I basically practiced my new knowledge of how overriding is done in C++.  It is similar to how it is done in Java, with a few differences in keywords and how interfaces and abstract classes are done in Java.  But the core concepts remain the same.  

In this exercise, the dog class Talk() function overrode the pure virtual Talk() function in the abstract Animal class.