Create main.cpp
This is one of the exercises I completed in the Udacity nano-degree programm.  In this example, I utilize the "override" keyword where a child class function is overriding the parent class interface (virtual) function or even a normal function.  This is different from Java, where the notion that a function is overriding another function is placed as a comment, rather than as code in C++.  

The exercise here also gives a hands on view of multiple inheritance.  The exercise in particular deals with particle and vehicular motion kinematics/dynamics when turning.