Create main.cpp
This is one of the exercises that I did as part of the Udacity nano-degree program I am enrolled in.  In this exercise I follow the template of class Dog, to create a class Cat that inherits from multiple parent classes, to learn how multiple inheritance is done in C++.  I also initialize and set the class for the assert tests.  

Multiple inheritance from multiple parent classes is something new to me.  In Java, I know that it is possible to "implement" multiple interfaces in a single class.  However in Java is not possible to inherit from a parent class, or "extend" in Java terminology, multiple parent classes in a single class.  This is different from what was seen in the C++ exercise.  In Java basically, one class can only extend only a single other class that is the parent class.  In C++ a class can be extended/inherit from multiple parent classes.