Create main.cpp
This is one of the exercises I had completed in the Udacity C++ nano-degree program I am enrolled in.  Here I am creating a template, which is used in generic programming.  Here the template creates a generic type called "Type".  Type can be of any type, in the example above, it is during the compile time, where the compiler will expand on any generic type functions (like Product() in the exercise) based on the type that is being returned or passed into the generic type function.