Create main.cpp
In this exercise that I completed, I created a generic function called average, by using a template.  At the same time, the code itself utilizes deduction, as there is no type specified when the average() generic function is called from main().  Deduction is when the compiler deduces the type of the template, by looking at what is being passed in the arguments to the generic function.