Create main.cpp
In this exercise, my task was to add just the Template declaration a class that utilizes two generic type attributes.  Before doing the exercise, there was no module that described how class declarations are to be done, and thus the purpose of this exercise was for me to think and utilize the knowledge that I have learned, to guess how this is to be done.  The purpose of the exercise was to be a challenge.  

I noticed that Mapping is called in a manner similar to how a generic function is called.  I also noticed in the constructor that the two generic attribute types are being used.  Thus I made a correct guess of placing the template declaration just before the class declaration/definition.  The two typenames were to obviously be KeyType and ValueType.  I checked to see if my response to the exercise worked, and it did.